Delete Risk.tsx and its related components:
- Delete `frontend/src/pages/Risk/Risk.tsx`
- Delete `frontend/src/pages/Risk/TopVulnerableDomains.tsx`
- Delete `frontend/src/pages/Risk/TopVulnerablePorts.tsx`
- Delete `frontend/src/pages/Risk/VSFindings.tsx`
- Delete `frontend/src/pages/Risk/VulnerabilityBarChart.tsx`
- Delete `frontend/src/pages/Risk/VulnerabilityCard.tsx`
- Delete `frontend/src/pages/Risk/WASFindings.tsx`
- Delete `frontend/src/pages/Risk/index.ts`
- Delete `frontend/src/pages/Risk/style.ts`
- Delete `frontend/src/pages/Risk/utils.ts`

- Relocate `getSeverityColor` utility function from `frontend/src/pages/Risk/utils.ts` to `frontend/src/utils/getSeverityColor.ts`.
- Update all imports of `getSeverityColor` in the following files to reflect the new location:
  - `frontend/src/pages/Domain/DomainDetails.tsx`
  - `frontend/src/pages/Domains/DomainDetails.tsx`
  - `frontend/src/pages/Vulnerabilities/Vulnerabilities.tsx`
  - `frontend/src/pages/Vulnerability/CveSection.tsx`
  - `frontend/src/pages/Vulnerability/OverviewSection.tsx`
