+++ frontend/src/pages/Domains/DomainDetails.tsx
@@ -21,7 +21,7 @@ import { DefinitionList } from 'components/DefinitionList';
 import { differenceInCalendarDays, parseISO } from 'date-fns';
 import { Webpage } from 'types';
 import { useAuthContext } from 'context';
-import { getSeverityColor } from 'pages/Risk/utils';
+import { getSeverityColor } from 'utils/getSeverityColor';
 import { Box } from '@mui/system';
 
 const PREFIX = 'DomainDetails';
