--- frontend/src/pages/Risk/Risk.module.scss
@@ -1,4 +1,0 @@
-.root {
-  max-width: 1500px;
-  margin: 0 auto;
-}
