--- frontend/src/pages/Risk/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Risk';
