+++ frontend/src/pages/Vulnerability/CveSection.tsx
@@ -12,7 +12,7 @@ import {
   TableRow
 } from '@mui/material';
 import { Cve } from 'types/cve';
-import { getSeverityColor } from 'pages/Risk/utils';
+import { getSeverityColor } from 'utils/getSeverityColor';
 
 interface CvssSubSectionProps {
   score?: string | null;
