+++ frontend/src/pages/Vulnerability/OverviewSection.tsx
@@ -11,7 +11,7 @@ import {
   Typography
 } from '@mui/material';
 import { Vulnerability as VulnerabilityType } from 'types';
-import { getSeverityColor } from 'pages/Risk/utils';
+import { getSeverityColor } from 'utils/getSeverityColor';
 import DetailsList from '@/components/Table/DetailsList';
 
 interface OverviewSectionProps {
